requirements.txt
openai
python-dotenv
flask[async]
flask-cors
Pillow